company_profiler/requirements.txt
openai
wikipedia-api
python-dotenv
